# [math] Rewrite the DARE solver to use a faster algorithm