--- a/math/test/discrete_algebraic_riccati_equation_test.cc
+++ b/math/test/discrete_algebraic_riccati_equation_test.cc
@@ -239,12 +239,14 @@ GTEST_TEST(DARETest, ACNotDetectable_ABQRN) {
 GTEST_TEST(DARETest, QDecomposition) {
   // Ensures the decomposition of Q into CᵀC is correct
 
-  const Eigen::Matrix2d A{{1.0, 0.0}, {0.0, 0.0}};
+  Eigen::Matrix2d A{2, 2};
+  A << 1.0, 0.0, 0.0, 0.0;
   const Eigen::Matrix2d B{Eigen::Matrix2d::Identity()};
   const Eigen::Matrix2d R{Eigen::Matrix2d::Identity()};
 
   // (A, C₁) should be detectable pair
-  const Eigen::Matrix2d C_1{{0.0, 0.0}, {1.0, 0.0}};
+  Eigen::Matrix2d C_1{2, 2};
+  C_1 << 0.0, 0.0, 1.0, 0.0;
   const Eigen::Matrix2d Q_1 = C_1.transpose() * C_1;
   EXPECT_NO_THROW(SolveDAREandVerify(A, B, Q_1, R));