Implements the SDA algorithm on page 5 of the following paper:

  E. K.-W. Chu, H.-Y. Fan, W.-W. Lin & C.-S. Wang "Structure-Preserving
  Algorithms for Periodic Discrete-Time Algebraic Riccati Equations",
  International Journal of Control, 77:8, 767-788, 2004.
  DOI: 10.1080/00207170410001714988

Here's Google Benchmark results for the old DARE solver using https://github.com/calcmogul/DAREBench/:
```
Running ./build/DAREBench
Run on (8 X 3595.49 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 6144 KiB (x1)
Load Average: 1.59, 4.28, 3.08
-------------------------------------------------------------------------------
Benchmark                                     Time             CPU   Iterations
-------------------------------------------------------------------------------
DARE_WPIMath_Dynamic                       28.2 us         28.1 us        24704
DARE_WPIMath_NoPrecondChecks_Dynamic       21.3 us         21.1 us        33349
DARE_WPIMath_Static                        13.1 us         13.1 us        52985
DARE_WPIMath_NoPrecondChecks_Static        8.78 us         8.77 us        79851
DARE_SLICOT                                41.4 us         41.3 us        16967
DARE_Drake                                 95.6 us         95.3 us         7597
```

Here's Google Benchmark results for the new DARE solver using https://github.com/calcmogul/DAREBench/tree/new-drake-dare-solver:
```
Running ./build/DAREBench
Run on (8 X 3600.03 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 6144 KiB (x1)
Load Average: 0.65, 0.41, 0.42
-------------------------------------------------------------------------------
Benchmark                                     Time             CPU   Iterations
-------------------------------------------------------------------------------
DARE_WPIMath_Dynamic                       28.1 us         28.1 us        25070
DARE_WPIMath_NoPrecondChecks_Dynamic       21.2 us         21.2 us        32991
DARE_WPIMath_Static                        13.1 us         13.1 us        52666
DARE_WPIMath_NoPrecondChecks_Static        8.79 us         8.79 us        79544
DARE_SLICOT                                41.3 us         41.3 us        16893
DARE_Drake                                 28.1 us         28.1 us        25109
```

"WPIMath" is a previous implementation I wrote of the new algorithm. Benchmarking has shown that using statically sized matrices is twice as fast as dynamically sized matrices, and the precondition checks are very expensive relative to the main solve.

WPIMath refactored the precondition checks out so users could opt out of precondition checking via separate DARE overloads, and so the ABQRN overload doesn't do an extra LLT decomposition of R. I decided to leave Drake's API as is though.

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/RobotLocomotion/drake/20104)
<!-- Reviewable:end -->