@RussTedrake 2023-08-31
@drake-jenkins-bot ok to test