@calcmogul 2023-08-31
Does Drake not use Eigen 3.4? That version added initializer list constructors for matrices, but the build is failing on them.