@jwnimmer-tri 2023-08-31
Most of our platforms have Eigen 3.4, but we still support Ubuntu 18.04 (for 6 more months), which only has 3.3.7.