@calcmogul 2023-08-31
Looks like cpplint is giving a false positive on braced initialization needing a space before the `{`.