@jwnimmer-tri 2023-08-31
> ... we still support Ubuntu 18.04 (for 6 more months), which only has 3.3.7.

Oops, I mis-spoke.  I should have said we still support Ubuntu **20.04**, which only has [3.3.7](https://packages.ubuntu.com/focal/libeigen3-dev).  We don't support 18.04.