@calcmogul 2023-08-31
The single test failure seems unrelated to the changes in this PR: https://drake-cdash.csail.mit.edu/test/1192182079