@jwnimmer-tri 2023-08-31
Yes, that's https://github.com/RobotLocomotion/drake/issues/19335 (sorry); you can ignore it.